Fixed reordering in Morphosyns.
Reordering (swapping) of words in Morphosyns was causing problems when
there were deleted words associated with the words to be swapped, for
example, “she” and “the” in “she wrote the letter” for the rule
swapping a verb and its object. This now works, but the conclusion is
really that swapping should be done in groups instead, as expensive as
that is.
Also fixed some details with how features are generated for English and
how they’re treated in Treetrans.build().